[Refactor] Ios 티어 탭, 지도 화면 뷰 포트 기반으로 성능 개선
