--- shared/appKit/src/iosMain/kotlin/com/kus/appkit/map/IosTierMapFactory.kt
@@ -1,14 +1,0 @@
-@file:OptIn(ExperimentalForeignApi::class)
-package com.kus.appkit.map
-
-import cocoapods.NMapsMap.NMFNaverMapView
-import com.kus.feature.tier.ui.map.TierMapFactory
-import com.kus.feature.tier.ui.map.TierMapHandle
-import kotlinx.cinterop.ExperimentalForeignApi
-
-class IosTierMapFactory : TierMapFactory {
-    override fun createMapHandle(): TierMapHandle {
-        val view = NMFNaverMapView()
-        return IosTierMapHandle(view)
-    }
-}
