+++ shared/appKit/src/iosMain/kotlin/com/kus/appkit/map/MapConstants.kt
@@ -1,0 +1,10 @@
+package com.kus.appkit.map
+
+
+const val SELECTED_TIER_MARKER_SCALE = 1.2
+
+const val DEFAULT_MARKER_WIDTH = 15.0
+const val DEFAULT_MARKER_HEIGHT = 20.0
+
+const val SAVED_MARKER_WIDTH = 12.0
+const val SAVED_MARKER_HEIGHT = 15.0
