+++ shared/appKit/src/iosMain/kotlin/com/kus/appkit/map/ViewportKey.kt
@@ -1,0 +1,8 @@
+package com.kus.appkit.map
+
+data class ViewportKey(
+    val south: Double,
+    val west: Double,
+    val north: Double,
+    val east: Double,
+)
